+++ src/Symfony/Component/Routing/Tests/Fixtures/dumper/compiled_url_matcher7.php
@@ -8,14 +8,14 @@
 return [
     false, // $matchHost
     [ // $staticRoutes
-        '/trailing/simple/no-methods' => [[['_route' => 'simple_trailing_slash_no_methods'], null, null, null, true, false, null]],
-        '/trailing/simple/get-method' => [[['_route' => 'simple_trailing_slash_GET_method'], null, ['GET' => 0], null, true, false, null]],
-        '/trailing/simple/head-method' => [[['_route' => 'simple_trailing_slash_HEAD_method'], null, ['HEAD' => 0], null, true, false, null]],
-        '/trailing/simple/post-method' => [[['_route' => 'simple_trailing_slash_POST_method'], null, ['POST' => 0], null, true, false, null]],
-        '/not-trailing/simple/no-methods' => [[['_route' => 'simple_not_trailing_slash_no_methods'], null, null, null, false, false, null]],
-        '/not-trailing/simple/get-method' => [[['_route' => 'simple_not_trailing_slash_GET_method'], null, ['GET' => 0], null, false, false, null]],
-        '/not-trailing/simple/head-method' => [[['_route' => 'simple_not_trailing_slash_HEAD_method'], null, ['HEAD' => 0], null, false, false, null]],
-        '/not-trailing/simple/post-method' => [[['_route' => 'simple_not_trailing_slash_POST_method'], null, ['POST' => 0], null, false, false, null]],
+        '/trailing/simple/no-methods' => [[['_route' => 'simple_trailing_slash_no_methods', '_route_path' => '/trailing/simple/no-methods/'], null, null, null, true, false, null]],
+        '/trailing/simple/get-method' => [[['_route' => 'simple_trailing_slash_GET_method', '_route_path' => '/trailing/simple/get-method/'], null, ['GET' => 0], null, true, false, null]],
+        '/trailing/simple/head-method' => [[['_route' => 'simple_trailing_slash_HEAD_method', '_route_path' => '/trailing/simple/head-method/'], null, ['HEAD' => 0], null, true, false, null]],
+        '/trailing/simple/post-method' => [[['_route' => 'simple_trailing_slash_POST_method', '_route_path' => '/trailing/simple/post-method/'], null, ['POST' => 0], null, true, false, null]],
+        '/not-trailing/simple/no-methods' => [[['_route' => 'simple_not_trailing_slash_no_methods', '_route_path' => '/not-trailing/simple/no-methods'], null, null, null, false, false, null]],
+        '/not-trailing/simple/get-method' => [[['_route' => 'simple_not_trailing_slash_GET_method', '_route_path' => '/not-trailing/simple/get-method'], null, ['GET' => 0], null, false, false, null]],
+        '/not-trailing/simple/head-method' => [[['_route' => 'simple_not_trailing_slash_HEAD_method', '_route_path' => '/not-trailing/simple/head-method'], null, ['HEAD' => 0], null, false, false, null]],
+        '/not-trailing/simple/post-method' => [[['_route' => 'simple_not_trailing_slash_POST_method', '_route_path' => '/not-trailing/simple/post-method'], null, ['POST' => 0], null, false, false, null]],
     ],
     [ // $regexpList
         0 => '{^(?'
@@ -34,15 +34,15 @@ return [
             .')/?$}sD',
     ],
     [ // $dynamicRoutes
-        46 => [[['_route' => 'regex_trailing_slash_no_methods'], ['param'], null, null, true, true, null]],
-        73 => [[['_route' => 'regex_trailing_slash_GET_method'], ['param'], ['GET' => 0], null, true, true, null]],
-        101 => [[['_route' => 'regex_trailing_slash_HEAD_method'], ['param'], ['HEAD' => 0], null, true, true, null]],
-        130 => [[['_route' => 'regex_trailing_slash_POST_method'], ['param'], ['POST' => 0], null, true, true, null]],
-        183 => [[['_route' => 'regex_not_trailing_slash_no_methods'], ['param'], null, null, false, true, null]],
-        211 => [[['_route' => 'regex_not_trailing_slash_GET_method'], ['param'], ['GET' => 0], null, false, true, null]],
-        240 => [[['_route' => 'regex_not_trailing_slash_HEAD_method'], ['param'], ['HEAD' => 0], null, false, true, null]],
+        46 => [[['_route' => 'regex_trailing_slash_no_methods', '_route_path' => '/trailing/regex/no-methods/{param}/'], ['param'], null, null, true, true, null]],
+        73 => [[['_route' => 'regex_trailing_slash_GET_method', '_route_path' => '/trailing/regex/get-method/{param}/'], ['param'], ['GET' => 0], null, true, true, null]],
+        101 => [[['_route' => 'regex_trailing_slash_HEAD_method', '_route_path' => '/trailing/regex/head-method/{param}/'], ['param'], ['HEAD' => 0], null, true, true, null]],
+        130 => [[['_route' => 'regex_trailing_slash_POST_method', '_route_path' => '/trailing/regex/post-method/{param}/'], ['param'], ['POST' => 0], null, true, true, null]],
+        183 => [[['_route' => 'regex_not_trailing_slash_no_methods', '_route_path' => '/not-trailing/regex/no-methods/{param}'], ['param'], null, null, false, true, null]],
+        211 => [[['_route' => 'regex_not_trailing_slash_GET_method', '_route_path' => '/not-trailing/regex/get-method/{param}'], ['param'], ['GET' => 0], null, false, true, null]],
+        240 => [[['_route' => 'regex_not_trailing_slash_HEAD_method', '_route_path' => '/not-trailing/regex/head-method/{param}'], ['param'], ['HEAD' => 0], null, false, true, null]],
         269 => [
-            [['_route' => 'regex_not_trailing_slash_POST_method'], ['param'], ['POST' => 0], null, false, true, null],
+            [['_route' => 'regex_not_trailing_slash_POST_method', '_route_path' => '/not-trailing/regex/post-method/{param}'], ['param'], ['POST' => 0], null, false, true, null],
             [null, null, null, null, false, false, 0],
         ],
     ],