+++ src/Symfony/Component/Routing/Tests/Fixtures/dumper/compiled_url_matcher9.php
@@ -9,9 +9,9 @@ return [
     true, // $matchHost
     [ // $staticRoutes
         '/' => [
-            [['_route' => 'a'], '{^(?P<d>[^\\.]++)\\.e\\.c\\.b\\.a$}sDi', null, null, false, false, null],
-            [['_route' => 'c'], '{^(?P<e>[^\\.]++)\\.e\\.c\\.b\\.a$}sDi', null, null, false, false, null],
-            [['_route' => 'b'], 'd.c.b.a', null, null, false, false, null],
+            [['_route' => 'a', '_route_path' => '/'], '{^(?P<d>[^\\.]++)\\.e\\.c\\.b\\.a$}sDi', null, null, false, false, null],
+            [['_route' => 'c', '_route_path' => '/'], '{^(?P<e>[^\\.]++)\\.e\\.c\\.b\\.a$}sDi', null, null, false, false, null],
+            [['_route' => 'b', '_route_path' => '/'], 'd.c.b.a', null, null, false, false, null],
         ],
     ],
     [ // $regexpList