+++ src/Symfony/Component/Routing/Tests/Matcher/RedirectableUrlMatcherTest.php
@@ -90,7 +90,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
             ->with('/foo/baz', 'foo', 'https')
             ->willReturn(['redirect' => 'value'])
         ;
-        $this->assertEquals(['_route' => 'foo', 'bar' => 'baz', 'redirect' => 'value'], $matcher->match('/foo/baz'));
+        $this->assertEquals(['_route' => 'foo', 'bar' => 'baz', 'redirect' => 'value', '_route_path' => '/foo/{bar}'], $matcher->match('/foo/baz'));
     }
 
     public function testSchemeRedirectForRoot()
@@ -104,7 +104,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
             ->method('redirect')
             ->with('/', 'foo', 'https')
             ->willReturn(['redirect' => 'value']);
-        $this->assertEquals(['_route' => 'foo', 'redirect' => 'value'], $matcher->match('/'));
+        $this->assertEquals(['_route' => 'foo', 'redirect' => 'value', '_route_path' => '/'], $matcher->match('/'));
     }
 
     public function testSlashRedirectWithParams()
@@ -119,7 +119,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
             ->with('/foo/baz/', 'foo', null)
             ->willReturn(['redirect' => 'value'])
         ;
-        $this->assertEquals(['_route' => 'foo', 'bar' => 'baz', 'redirect' => 'value'], $matcher->match('/foo/baz'));
+        $this->assertEquals(['_route' => 'foo', 'bar' => 'baz', 'redirect' => 'value', '_route_path' => '/foo/{bar}/'], $matcher->match('/foo/baz'));
     }
 
     public function testRedirectPreservesUrlEncoding()
@@ -138,7 +138,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
         $coll->add('foo', new Route('/foo', [], [], [], '', ['https']));
         $matcher = $this->getUrlMatcher($coll, new RequestContext());
         $matcher->expects($this->once())->method('redirect')->with('/foo', 'foo', 'https')->willReturn([]);
-        $this->assertSame(['_route' => 'foo'], $matcher->match('/foo'));
+        $this->assertSame(['_route' => 'foo', '_route_path' => '/foo'], $matcher->match('/foo'));
     }
 
     public function testFallbackPage()
@@ -149,7 +149,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
 
         $matcher = $this->getUrlMatcher($coll);
         $matcher->expects($this->once())->method('redirect')->with('/foo/', 'foo')->willReturn(['_route' => 'foo']);
-        $this->assertSame(['_route' => 'foo'], $matcher->match('/foo'));
+        $this->assertSame(['_route' => 'foo', '_route_path' => '/foo/'], $matcher->match('/foo'));
 
         $coll = new RouteCollection();
         $coll->add('foo', new Route('/foo'));
@@ -157,7 +157,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
 
         $matcher = $this->getUrlMatcher($coll);
         $matcher->expects($this->once())->method('redirect')->with('/foo', 'foo')->willReturn(['_route' => 'foo']);
-        $this->assertSame(['_route' => 'foo'], $matcher->match('/foo/'));
+        $this->assertSame(['_route' => 'foo', '_route_path' => '/foo'], $matcher->match('/foo/'));
     }
 
     public function testMissingTrailingSlashAndScheme()
@@ -180,6 +180,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
         $expected = [
             '_route' => 'b',
             'customerId' => '123',
+            '_route_path' => '/api/customers/{customerId}/contactpersons/',
         ];
         $this->assertEquals($expected, $matcher->match('/api/customers/123/contactpersons/'));
 
@@ -195,7 +196,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
         $matcher = $this->getUrlMatcher($coll);
         $matcher->expects($this->once())->method('redirect')->with('/123')->willReturn([]);
 
-        $this->assertEquals(['_route' => 'a', 'a' => '123'], $matcher->match('/123/'));
+        $this->assertEquals(['_route' => 'a', 'a' => '123', '_route_path' => '/{a}'], $matcher->match('/123/'));
     }
 
     public function testTrailingRequirementWithDefaultA()
@@ -206,7 +207,7 @@ class RedirectableUrlMatcherTest extends UrlMatcherTest
         $matcher = $this->getUrlMatcher($coll);
         $matcher->expects($this->once())->method('redirect')->with('/fr-fr')->willReturn([]);
 
-        $this->assertEquals(['_route' => 'a', 'a' => 'aaa'], $matcher->match('/fr-fr/'));
+        $this->assertEquals(['_route' => 'a', 'a' => 'aaa', '_route_path' => '/fr-fr/{a}'], $matcher->match('/fr-fr/'));
     }
 
     protected function getUrlMatcher(RouteCollection $routes, RequestContext $context = null)